moved load/write functions for high-level features to separate file.
