+++ bin/BeatTracker
@@ -95,7 +95,7 @@ def main():
         # track the beats in the activation function
         beat_processor = BeatTrackingProcessor(**vars(args))
         # output handler
-        from madmom.utils import write_events as writer
+        from madmom.io import write_events as writer
         # sequentially process them
         out_processor = [beat_processor, writer]
