+++ bin/CRFBeatDetector
@@ -97,7 +97,7 @@ def main():
         # detect the beats with a CRF
         beat_processor = CRFBeatDetectionProcessor(**vars(args))
         # output handler
-        from madmom.utils import write_events as writer
+        from madmom.io import write_events as writer
         # sequentially process them
         out_processor = [beat_processor, writer]
