+++ bin/PianoTranscriptor
@@ -13,8 +13,8 @@ from madmom.processors import IOProcessor, io_arguments
 from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.notes import (RNNPianoNoteProcessor,
-                                   NotePeakPickingProcessor, write_midi,
-                                   write_notes, write_mirex_format)
+                                   NotePeakPickingProcessor)
+from madmom.io import write_notes, write_notes_midi, write_notes_mirex
 
 
 def main():
@@ -105,9 +105,9 @@ def main():
         if args.output_format is None:
             output = write_notes
         elif args.output_format == 'midi':
-            output = write_midi
+            output = write_notes_midi
         elif args.output_format == 'mirex':
-            output = write_mirex_format
+            output = write_notes_mirex
         else:
             raise ValueError('unknown output format: %s' % args.output_format)
         out_processor = [peak_picking, output]
