+++ bin/SpectralOnsetDetection
@@ -105,7 +105,7 @@ def main():
     else:
         # perform peak picking of the onset function
         peak_picking = OnsetPeakPickingProcessor(**vars(args))
-        from madmom.utils import write_events as writer
+        from madmom.io import write_events as writer
         # sequentially process them
         out_processor = [peak_picking, writer]
