+++ bin/TempoDetector
@@ -12,7 +12,8 @@ from madmom.processors import IOProcessor, io_arguments
 from madmom.audio.signal import SignalProcessor
 from madmom.features import ActivationsProcessor
 from madmom.features.beats import RNNBeatProcessor
-from madmom.features.tempo import TempoEstimationProcessor, write_tempo
+from madmom.features.tempo import TempoEstimationProcessor
+from madmom.io import write_tempo, write_events as writer
 
 
 def main():
@@ -97,7 +98,7 @@ def main():
             writer = partial(write_tempo, mirex=True)
         elif args.tempo_format in ('raw', 'all'):
             # borrow the event writer for outputting multiple values
-            from madmom.utils import write_events as writer
+            pass
         else:
             # normal output
             writer = write_tempo
