+++ madmom/utils/__init__.py
@@ -256,69 +256,6 @@ def match_file(filename, match_list, suffix=None, match_suffix=None,
     return matches
 
 
-@suppress_warnings
-def load_events(filename):
-    """
-    Load a events from a text file, one floating point number per line.
-
-    Parameters
-    ----------
-    filename : str or file handle
-        File to load the events from.
-
-    Returns
-    -------
-    numpy array
-        Events.
-
-    Notes
-    -----
-    Comments (lines starting with '#') and additional columns are ignored,
-    i.e. only the first column is returned.
-
-    """
-    # read in the events, one per line
-    events = np.loadtxt(filename, ndmin=2)
-    # 1st column is the event's time, the rest is ignored
-    return events[:, 0]
-
-
-def write_events(events, filename, fmt='%.3f', delimiter='\t', header=''):
-    """
-    Write events to a text file, one event per line.
-
-    Parameters
-    ----------
-    events : numpy array
-        Events to be written to file.
-    filename : str or file handle
-        File to write the events to.
-    fmt : str, optional
-        How to format the events.
-    delimiter : str, optional
-        String or character separating multiple columns.
-    header : str, optional
-        Header to be written (as a comment).
-
-    Returns
-    -------
-    numpy array
-        Events.
-
-    Notes
-    -----
-    This function is just a wrapper to ``np.savetxt``, but reorders the
-    arguments in a way it can be used as an
-    :class:`.processors.OutputProcessor`.
-
-    """
-    # write the events to the output
-    np.savetxt(filename, np.asarray(events),
-               fmt=fmt, delimiter=delimiter, header=header)
-    # also return them
-    return events
-
-
 def combine_events(events, delta, combine='mean'):
     """
     Combine all events within a certain range.
